backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.0
supabase==2.3.5
twilio==8.10.0
groq==0.4.1
pydantic==2.5.0
python-multipart==0.0.6
